+++ src/train_randomfs.m
@@ -185,11 +185,11 @@ for i = 1:length(ind_samples)
     landmark = Tr_Data{s}.intermediate_shapes{stage}(lmarkID, :, k);
     [mean_landmark_x, mean_landmark_y] = tformfwd(Tr_Data{s}.tf2meanshape{k}, landmark(1), landmark(2));
         
-    pixel_a_x = int16(bsxfun(@plus, pixel_a_x_lmcoord, mean_landmark_x));
-    pixel_a_y = int16(bsxfun(@plus, pixel_a_y_lmcoord, mean_landmark_y));
+    pixel_a_x = (bsxfun(@plus, pixel_a_x_lmcoord, mean_landmark_x));
+    pixel_a_y = (bsxfun(@plus, pixel_a_y_lmcoord, mean_landmark_y));
     
-    pixel_b_x = int16(bsxfun(@plus, pixel_b_x_lmcoord, mean_landmark_x));
-    pixel_b_y = int16(bsxfun(@plus, pixel_b_y_lmcoord, mean_landmark_y));
+    pixel_b_x = (bsxfun(@plus, pixel_b_x_lmcoord, mean_landmark_x));
+    pixel_b_y = (bsxfun(@plus, pixel_b_y_lmcoord, mean_landmark_y));
     
     % transform the pixels from image coordinate (meanshape) to coordinate of current shape
     
@@ -199,11 +199,11 @@ for i = 1:length(ind_samples)
     width = (Tr_Data{s}.width);
     height = (Tr_Data{s}.height);
 
-    pixel_a_x = max(1, min(pixel_a_x, width));
-    pixel_a_y = max(1, min(pixel_a_y, height));
-    
-    pixel_b_x = max(1, min(pixel_b_x, width));
-    pixel_b_y = max(1, min(pixel_b_y, height));
+    pixel_a_x = max(1, min(int16(pixel_a_x), width));
+    pixel_a_y = max(1, min(int16(pixel_a_y), height));
+
+    pixel_b_x = max(1, min(int16(pixel_b_x), width));
+    pixel_b_y = max(1, min(int16(pixel_b_y), height));
     
     pdfeats(:, i) = double(Tr_Data{s}.img_gray(pixel_a_y + (pixel_a_x-1)*height)) - double(Tr_Data{s}.img_gray(pixel_b_y + (pixel_b_x-1)*height));
